renamed abstract-operators to get-operators
