+++ packages/@orbit/jsonapi/src/lib/query-operators.ts
@@ -4,7 +4,7 @@ import {
 } from '@orbit/data';
 import JSONAPISource from '../jsonapi-source';
 import { JSONAPIDocument } from '../jsonapi-document';
-import { AbstractOperators } from "./abstract-operators";
+import { GetOperators } from "./get-operators";
 import { DeserializedDocument } from "../jsonapi-serializer";
 
 function deserialize(source: JSONAPISource, document: JSONAPIDocument): DeserializedDocument {
@@ -17,22 +17,22 @@ export interface QueryOperator {
 
 export const QueryOperators: Dict<QueryOperator> = {
   findRecord(source: JSONAPISource, query: Query) {
-    return AbstractOperators.findRecord(source, query)
+    return GetOperators.findRecord(source, query)
       .then(data => deserialize(source, data));
   },
 
   findRecords(source: JSONAPISource, query: Query) {
-    return AbstractOperators.findRecords(source, query)
+    return GetOperators.findRecords(source, query)
       .then(data => deserialize(source, data));
   },
 
   findRelatedRecord(source: JSONAPISource, query: Query) {
-    return AbstractOperators.findRelatedRecord(source, query)
+    return GetOperators.findRelatedRecord(source, query)
       .then(data => deserialize(source, data));
   },
 
   findRelatedRecords(source: JSONAPISource, query: Query) {
-    return AbstractOperators.findRelatedRecords(source, query)
+    return GetOperators.findRelatedRecords(source, query)
       .then(data => deserialize(source, data));
   }
 };
